+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/rings/RingOfMight.java
@@ -94,7 +94,7 @@ public class RingOfMight extends Ring {
 	@Override
 	public String upgradeStat1(long level) {
 		if (cursed && cursedKnown) level = Math.min(-1, level-3);
-		return Integer.toString(level+1);
+		return Long.toString(level+1);
 	}
 
 	@Override
