+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/AssassinsBlade.java
@@ -93,7 +93,7 @@ public class AssassinsBlade extends MeleeWeapon {
 
 	@Override
 	public String upgradeAbilityStat(long level) {
-		return Integer.toString(2+level);
+		return Long.toString(2+level);
 	}
 
 }
