+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/Longsword.java
@@ -264,7 +264,7 @@ public class Longsword extends MeleeWeapon {
 	}
 
 	public String upgradeAbilityStat(long level){
-		int dmgBoost = 6 + level;
+		long dmgBoost = 6 + level;
 		return augment.damageFactor(min(level)+dmgBoost) + "-" + augment.damageFactor(max(level)+dmgBoost);
 	}
 }
