+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/melee/Quarterstaff.java
@@ -75,7 +75,7 @@ public class Quarterstaff extends MeleeWeapon {
 
 	@Override
 	public String upgradeAbilityStat(long level) {
-		return Integer.toString(4+level);
+		return Long.toString(4+level);
 	}
 
 	public static class DefensiveStance extends FlavourBuff {
